storage: consolidate RawStatusUpdate into StatusUpdate impls
RawStatusUpdate had become nothing but a holder for impls on
StatusUpdate. Instead, move those impls onto the parent struct.
