+++ src/storage-controller/src/healthcheck.rs
@@ -7,15 +7,15 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use chrono::{DateTime, Utc};
 use differential_dataflow::lattice::Lattice;
 use mz_ore::now::EpochMillis;
 use mz_persist_types::Codec64;
 use mz_repr::{ColumnName, Datum, GlobalId, Row, TimestampManipulation};
+use mz_storage_client::client::StatusUpdate;
 use mz_storage_client::healthcheck;
 use once_cell::sync::Lazy;
 use timely::progress::Timestamp;
@@ -24,91 +24,6 @@ use crate::collection_mgmt::CollectionManager;
 use crate::IntrospectionType;
 
 pub use mz_storage_client::healthcheck::*;
-
-/// Represents an entry in `mz_internal.mz_{source|sink}_status_history`
-/// For brevity's sake, this is represented as a single type because currently
-/// the two relation's have the same shape. If thta changes, we can bifurcate this
-#[derive(Clone)]
-pub struct RawStatusUpdate {
-    /// The ID of the collection
-    pub id: GlobalId,
-    /// The timestamp of the status update
-    pub ts: DateTime<Utc>,
-    /// The name of the status update
-    pub status_name: String,
-    /// A succinct error, if any
-    pub error: Option<String>,
-    /// Hints about the error, if any
-    pub hints: BTreeSet<String>,
-    /// The full set of namespaced errors that consolidated into the `error`, if any
-    pub namespaced_errors: BTreeMap<String, String>,
-}
-
-impl RawStatusUpdate {
-    pub fn dropped_status(id: GlobalId, ts: DateTime<Utc>) -> RawStatusUpdate {
-        RawStatusUpdate {
-            id,
-            ts,
-            status_name: "dropped".to_string(),
-            error: None,
-            hints: Default::default(),
-            namespaced_errors: Default::default(),
-        }
-    }
-
-    fn produce_new_status(prev: &str, new: &str) -> bool {
-        match (prev, new) {
-            // TODO(guswynn): Ideally only `failed` sources should not be marked as paused.
-            // Additionally, dropping a replica and then restarting environmentd will
-            // fail this check. This will all be resolved in:
-            // https://github.com/MaterializeInc/materialize/pull/23013
-            (prev, new) if prev == "stalled" && new == "paused" => false,
-            // Don't re-mark that object as paused.
-            (prev, new) if prev == "paused" && new == "paused" => false,
-            // De-duplication of other statuses is currently managed by the
-            // `health_operator`.
-            _ => true,
-        }
-    }
-}
-
-impl From<RawStatusUpdate> for Row {
-    fn from(update: RawStatusUpdate) -> Self {
-        let timestamp = Datum::TimestampTz(update.ts.try_into().expect("must fit"));
-        let id = update.id.to_string();
-        let id = Datum::String(&id);
-        let status = Datum::String(&update.status_name);
-        let error = update.error.as_deref().into();
-
-        let mut row = Row::default();
-        let mut packer = row.packer();
-        packer.extend([timestamp, id, status, error]);
-
-        if !update.hints.is_empty() || !update.namespaced_errors.is_empty() {
-            packer.push_dict_with(|dict_packer| {
-                // `hint` and `namespaced` are ordered,
-                // as well as the BTree's they each contain.
-                if !update.hints.is_empty() {
-                    dict_packer.push(Datum::String("hints"));
-                    dict_packer.push_list(update.hints.iter().map(|s| Datum::String(s)));
-                }
-                if !update.namespaced_errors.is_empty() {
-                    dict_packer.push(Datum::String("namespaced"));
-                    dict_packer.push_dict(
-                        update
-                            .namespaced_errors
-                            .iter()
-                            .map(|(k, v)| (k.as_str(), Datum::String(v))),
-                    );
-                }
-            });
-        } else {
-            packer.push(Datum::Null);
-        }
-
-        row
-    }
-}
 
 /// A lightweight wrapper around [`CollectionManager`] that assists with
 /// appending status updates to ntrospection statuses.
@@ -175,8 +90,8 @@ static SOURCE_STATUS_UPDATES: Lazy<IntrospectionStatusUpdate> = Lazy::new(|| {
             let new = new.unwrap_str();
             let prev = prev.unwrap_str();
 
-            // The rows for SourceStatusUpdates are derived from RawStatusUpdate
-            RawStatusUpdate::produce_new_status(prev, new)
+            // The rows for SourceStatusUpdates are derived from StatusUpdate
+            StatusUpdate::new_status_succeeds_prev(prev, new)
         },
     }
 });
@@ -197,8 +112,8 @@ static SINK_STATUS_UPDATES: Lazy<IntrospectionStatusUpdate> = Lazy::new(|| {
             let new = new.unwrap_str();
             let prev = prev.unwrap_str();
 
-            // The rows for SinkStatusUpdates are derived from RawStatusUpdate
-            RawStatusUpdate::produce_new_status(prev, new)
+            // The rows for SourceStatusUpdates are derived from StatusUpdate
+            StatusUpdate::new_status_succeeds_prev(prev, new)
         },
     }
 });
@@ -339,6 +254,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::collections::BTreeSet;
 
     #[mz_ore::test]
     fn test_row() {
@@ -346,10 +262,10 @@ mod tests {
         let hint = "hint message";
         let id = GlobalId::User(1);
         let status = "dropped";
-        let row = Row::from(RawStatusUpdate {
+        let row = Row::from(StatusUpdate {
             id,
-            ts: chrono::offset::Utc::now(),
-            status_name: status.to_string(),
+            timestamp: chrono::offset::Utc::now(),
+            status: status.to_string(),
             error: Some(error_message.to_string()),
             hints: BTreeSet::from([hint.to_string()]),
             namespaced_errors: Default::default(),
@@ -390,10 +306,10 @@ mod tests {
         let error_message = "error message";
         let id = GlobalId::User(1);
         let status = "dropped";
-        let row = Row::from(RawStatusUpdate {
+        let row = Row::from(StatusUpdate {
             id,
-            ts: chrono::offset::Utc::now(),
-            status_name: status.to_string(),
+            timestamp: chrono::offset::Utc::now(),
+            status: status.to_string(),
             error: Some(error_message.to_string()),
             hints: Default::default(),
             namespaced_errors: Default::default(),
@@ -418,10 +334,10 @@ mod tests {
         let id = GlobalId::User(1);
         let status = "dropped";
         let hint = "hint message";
-        let row = Row::from(RawStatusUpdate {
+        let row = Row::from(StatusUpdate {
             id,
-            ts: chrono::offset::Utc::now(),
-            status_name: status.to_string(),
+            timestamp: chrono::offset::Utc::now(),
+            status: status.to_string(),
             error: None,
             hints: BTreeSet::from([hint.to_string()]),
             namespaced_errors: Default::default(),
@@ -462,10 +378,10 @@ mod tests {
         let error_message = "error message";
         let id = GlobalId::User(1);
         let status = "dropped";
-        let row = Row::from(RawStatusUpdate {
+        let row = Row::from(StatusUpdate {
             id,
-            ts: chrono::offset::Utc::now(),
-            status_name: status.to_string(),
+            timestamp: chrono::offset::Utc::now(),
+            status: status.to_string(),
             error: Some(error_message.to_string()),
             hints: Default::default(),
             namespaced_errors: BTreeMap::from([("thing".to_string(), "error".to_string())]),
@@ -507,10 +423,10 @@ mod tests {
         let hint = "hint message";
         let id = GlobalId::User(1);
         let status = "dropped";
-        let row = Row::from(RawStatusUpdate {
+        let row = Row::from(StatusUpdate {
             id,
-            ts: chrono::offset::Utc::now(),
-            status_name: status.to_string(),
+            timestamp: chrono::offset::Utc::now(),
+            status: status.to_string(),
             error: Some(error_message.to_string()),
             hints: BTreeSet::from([hint.to_string()]),
             namespaced_errors: BTreeMap::from([("thing".to_string(), "error".to_string())]),
